+++ packages/rich-text-editor/src/RichTextEditor/schema.ts
@@ -1,0 +1,51 @@
+import { MarkSpec, NodeSpec, Schema } from "prosemirror-model"
+import boldIcon from "@kaizen/component-library/icons/bold.icon.svg"
+import italicIcon from "@kaizen/component-library/icons/italics.icon.svg"
+import underlineIcon from "@kaizen/component-library/icons/underline.icon.svg"
+import {
+  nodes as coreNodes,
+  marks as coreMarks,
+} from "@cultureamp/rich-text-toolkit"
+
+export const nodes: NodeSpec = {
+  doc: coreNodes.doc,
+  paragraph: coreNodes.paragraph,
+  text: coreNodes.text,
+  // eslint-disable-next-line camelcase
+  hard_break: coreNodes.hard_break,
+}
+
+export const marks: MarkSpec = {
+  strong: {
+    ...coreMarks.strong,
+    control: {
+      label: "Bold",
+      icon: boldIcon,
+    },
+  },
+  em: {
+    ...coreMarks.em,
+    control: {
+      label: "Italic",
+      icon: italicIcon,
+    },
+  },
+  underline: {
+    ...coreMarks.underline,
+    control: {
+      label: "Underline",
+      icon: underlineIcon,
+    },
+  },
+}
+
+export const createSchemaFromControls = controls => {
+  const newMarks: MarkSpec = controls.reduce(
+    (previousValue, currentValue) => ({
+      ...previousValue,
+      [currentValue]: marks[currentValue],
+    }),
+    {}
+  )
+  return new Schema({ nodes, marks: newMarks })
+}