+++ packages/rich-text-editor/src/RichTextEditor/types.ts
@@ -1,0 +1,1 @@
+export type EditorContentArray = Array<{ [key: string]: any }>