+++ packages/component-library/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [13.2.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/component-library@13.2.5...@kaizen/component-library@13.2.6) (2022-04-03)
+
+**Note:** Version bump only for package @kaizen/component-library
+
+
+
+
+
 ## [13.2.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/component-library@13.2.4...@kaizen/component-library@13.2.5) (2022-03-31)