+++ packages/rich-text-editor/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.4.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/rich-text-editor@1.3.1...@kaizen/rich-text-editor@1.4.0) (2022-04-04)
+
+
+### Features
+
+* Add inputrules to RTE ([#2667](https://github.com/cultureamp/kaizen-design-system/issues/2667)) ([4831889](https://github.com/cultureamp/kaizen-design-system/commit/4831889d8edb42e39109fa524fad72d009296f68))
+
+
+
+
+
+## [1.3.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/rich-text-editor@1.3.0...@kaizen/rich-text-editor@1.3.1) (2022-04-04)
+
+
+### Bug Fixes
+
+* Set pointer-events to none on RTE focus ring ([#2669](https://github.com/cultureamp/kaizen-design-system/issues/2669)) ([0252656](https://github.com/cultureamp/kaizen-design-system/commit/0252656029f0cebec77f7263c528bba45ce61b07))
+
+
+
+
+
+# [1.3.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/rich-text-editor@1.2.4...@kaizen/rich-text-editor@1.3.0) (2022-04-04)
+
+
+### Features
+
+* Create RichTextEditor component (sans controls) ([#2650](https://github.com/cultureamp/kaizen-design-system/issues/2650)) ([3f9111a](https://github.com/cultureamp/kaizen-design-system/commit/3f9111aee9c452fb0cc5e4fc2c1295de36d80ae2))
+
+
+
+
+
+## [1.2.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/rich-text-editor@1.2.3...@kaizen/rich-text-editor@1.2.4) (2022-04-03)
+
+**Note:** Version bump only for package @kaizen/rich-text-editor
+
+
+
+
+
 ## [1.2.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/rich-text-editor@1.2.2...@kaizen/rich-text-editor@1.2.3) (2022-03-31)